Look for alteration based on protein change and name
This solves issue #1284
The issue is when we looking for alteration, we compare the protein change only. When we try to update the display name, since you always get the same alteration then you will use the old one assigned to new evidence which the new alteration should be created.