Group relevant alterations based on alteration without exclusion criteria
When alteration has exclusion criteria and however is included in relevantAlterations, that means the alteration in query is not excluded. Therefore the relevant alteration should be treated as the same to the one without exclusion criteria. For instance, Oncogenic Mutations and Oncogenic Mutations {excluding V600K} as relevant alterations when alteration in query is V600E, both relevant alterations are the same.